requirements.txt
# Common requirements
build>=1.2.1
setuptools==67.8.0
wheel==0.38.4
requests==2.29.0
python-decouple==3.8
typing-extensions~=4.12.2

# Tests requirements
requests-mock>=1.12.1
pyjwt>=2.8.0
pytest>=7.4.0
coverage==6.4.4
flake8>=4.0.1
